requirements.txt
fastapi
uvicorn[standard]
playwright
python-dotenv
pyotp
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
httpx
